requirements.txt
scipy
yfinance
pandas
numpy
PyYAML
